+++ lista.h
@@ -21,13 +21,13 @@
     bool Lista_Cheia(LISTA *Lista);
     bool Lista_Vazia(LISTA *Lista);
 
-    bool Lista_Inserir(LISTA *Lista, int zeros, int valor, int categoria);
+    bool Lista_Inserir(LISTA *Lista, int zeros, float valor, int categoria);
 
     NO *Lista_No(LISTA *Lista, int index);
     NO *Lista_No_Proximo(NO *No);
 
     int Lista_No_Zeros(NO *No);
-    int Lista_No_Valor(NO *No);
+    float Lista_No_Valor(NO *No);
     int Lista_No_Categoria(NO *No);
 
 #endif
